+++ java/client/src/main/java/glide/api/commands/SortedSetBaseCommands.java
@@ -12,7 +12,7 @@ import glide.api.models.commands.RangeOptions.RangeQuery;
 import glide.api.models.commands.RangeOptions.ScoreBoundary;
 import glide.api.models.commands.RangeOptions.ScoreRange;
 import glide.api.models.commands.RangeOptions.ScoredRangeQuery;
-import glide.api.models.commands.ScoreModifier;
+import glide.api.models.commands.ScoreFilter;
 import glide.api.models.commands.WeightAggregateOptions.Aggregate;
 import glide.api.models.commands.WeightAggregateOptions.KeyArray;
 import glide.api.models.commands.WeightAggregateOptions.KeysOrWeightedKeys;
@@ -958,7 +958,7 @@ public interface SortedSetBaseCommands {
     /**
      * Blocks the connection until it pops and returns a member-score pair from the first non-empty
      * sorted set, with the given <code>keys</code> being checked in the order they are provided.<br>
-     * To pop more than one element use {@link #bzmpop(String[], ScoreModifier, double, long)}.<br>
+     * To pop more than one element use {@link #bzmpop(String[], ScoreFilter, double, long)}.<br>
      * <code>BZMPOP</code> is the blocking variant of <code>ZMPOP</code>.
      *
      * @apiNote
@@ -973,8 +973,8 @@ public interface SortedSetBaseCommands {
      * @since Redis 7.0 and above
      * @see <a href="https://redis.io/commands/bzmpop/">redis.io</a> for more details.
      * @param keys The keys of the sorted sets.
-     * @param modifier The element pop criteria - either {@link ScoreModifier#MIN} or {@link
-     *     ScoreModifier#MAX} to pop members with the lowest/highest scores accordingly.
+     * @param modifier The element pop criteria - either {@link ScoreFilter#MIN} or {@link
+     *     ScoreFilter#MAX} to pop members with the lowest/highest scores accordingly.
      * @param timeout The number of seconds to wait for a blocking operation to complete. A value of
      *     <code>0</code> will block indefinitely.
      * @return A two-element <code>array</code> containing the key name of the set from which an
@@ -988,7 +988,7 @@ public interface SortedSetBaseCommands {
      * System.out.printf("Popped '%s' with score %d from '%s'%n", element, data.get(element), result[0]);
      * }</pre>
      */
-    CompletableFuture<Object[]> bzmpop(String[] keys, ScoreModifier modifier, double timeout);
+    CompletableFuture<Object[]> bzmpop(String[] keys, ScoreFilter modifier, double timeout);
 
     // TODO add @link to ZMPOP when implemented
     /**
@@ -1009,8 +1009,8 @@ public interface SortedSetBaseCommands {
      * @since Redis 7.0 and above
      * @see <a href="https://redis.io/commands/bzmpop/">redis.io</a> for more details.
      * @param keys The keys of the sorted sets.
-     * @param modifier The element pop criteria - either {@link ScoreModifier#MIN} or {@link
-     *     ScoreModifier#MAX} to pop members with the lowest/highest scores accordingly.
+     * @param modifier The element pop criteria - either {@link ScoreFilter#MIN} or {@link
+     *     ScoreFilter#MAX} to pop members with the lowest/highest scores accordingly.
      * @param timeout The number of seconds to wait for a blocking operation to complete. A value of
      *     <code>0</code> will block indefinitely.
      * @param count The number of elements to pop.
@@ -1027,7 +1027,7 @@ public interface SortedSetBaseCommands {
      * }</pre>
      */
     CompletableFuture<Object[]> bzmpop(
-            String[] keys, ScoreModifier modifier, double timeout, long count);
+            String[] keys, ScoreFilter modifier, double timeout, long count);
 
     /**
      * Returns the union of members from sorted sets specified by the given <code>keysOrWeightedKeys
