+++ java/integTest/src/test/java/glide/TransactionTestUtilities.java
@@ -5,8 +5,8 @@ import static glide.TestConfiguration.REDIS_VERSION;
 import static glide.api.BaseClient.OK;
 import static glide.api.models.commands.FlushMode.ASYNC;
 import static glide.api.models.commands.LInsertOptions.InsertPosition.AFTER;
-import static glide.api.models.commands.ScoreModifier.MAX;
-import static glide.api.models.commands.ScoreModifier.MIN;
+import static glide.api.models.commands.ScoreFilter.MAX;
+import static glide.api.models.commands.ScoreFilter.MIN;
 import static glide.utils.ArrayTransformUtils.concatenateArrays;
 
 import glide.api.models.BaseTransaction;
